Use long for price instead of decimal Add Price in utilities

+++ Entities/Booking/Order.cs
@@ -27,7 +27,7 @@ namespace PEAS.Entities.Booking
 
         public required Currency Currency { get; set; }
 
-        public required decimal Price { get; set; }
+        public required long Price { get; set; }
 
         public required string Title { get; set; }
